# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "新能源监控系统RESTful API接口文档，提供区域管理、厂站管理、设备管理、采集点管理、告警管理、数据查询等功能。",
        "title": "新能源监控系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API支持团队",
            "url": "http://www.new-energy-monitoring.com/support",
            "email": "support@new-energy-monitoring.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/devices": {
            "get": {
                "description": "获取所有设备的列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "获取设备列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "厂站ID",
                        "name": "station_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PagedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.DeviceResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "创建设备",
                "parameters": [
                    {
                        "description": "设备信息",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DeviceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "description": "根据ID获取设备详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "获取设备详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DeviceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新设备信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "更新设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设备信息",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.DeviceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "设备管理"
                ],
                "summary": "删除设备",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateDeviceRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "station_id",
                "type"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INV_001"
                },
                "description": {
                    "type": "string",
                    "example": "逆变器设备"
                },
                "install_date": {
                    "type": "string",
                    "example": "2024-01-01"
                },
                "ip_address": {
                    "type": "string",
                    "example": "192.168.1.101"
                },
                "manufacturer": {
                    "type": "string",
                    "example": "华为"
                },
                "model": {
                    "type": "string",
                    "example": "SUN2000-100KTL"
                },
                "name": {
                    "type": "string",
                    "example": "1号逆变器"
                },
                "port": {
                    "type": "integer",
                    "example": 502
                },
                "protocol": {
                    "type": "string",
                    "example": "modbus"
                },
                "rated_current": {
                    "type": "number",
                    "example": 150
                },
                "rated_power": {
                    "type": "number",
                    "example": 100
                },
                "rated_voltage": {
                    "type": "number",
                    "example": 380
                },
                "serial_number": {
                    "type": "string",
                    "example": "SN123456"
                },
                "slave_id": {
                    "type": "integer",
                    "example": 1
                },
                "station_id": {
                    "type": "string",
                    "example": "station-001"
                },
                "type": {
                    "type": "string",
                    "example": "inverter"
                },
                "warranty_date": {
                    "type": "string",
                    "example": "2025-01-01"
                }
            }
        },
...
        "dto.DeviceResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INV_001"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "逆变器设备"
                },
                "id": {
                    "type": "string",
                    "example": "device-001"
                },
                "install_date": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string",
                    "example": "192.168.1.101"
                },
                "last_online": {
                    "type": "string"
                },
                "manufacturer": {
                    "type": "string",
                    "example": "华为"
                },
                "model": {
                    "type": "string",
                    "example": "SUN2000-100KTL"
                },
                "name": {
                    "type": "string",
                    "example": "1号逆变器"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PointBrief"
                    }
                },
                "port": {
                    "type": "integer",
                    "example": 502
                },
                "protocol": {
                    "type": "string",
                    "example": "modbus"
                },
                "rated_current": {
                    "type": "number",
                    "example": 150
                },
                "rated_power": {
                    "type": "number",
                    "example": 100
                },
                "rated_voltage": {
                    "type": "number",
                    "example": 380
                },
                "serial_number": {
                    "type": "string",
                    "example": "SN123456"
                },
                "slave_id": {
                    "type": "integer",
                    "example": 1
                },
                "station_id": {
                    "type": "string",
                    "example": "station-001"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "example": "inverter"
                },
                "updated_at": {
                    "type": "string"
                },
                "warranty_date": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1709500800000
                }
            }
        },
...
        "dto.PagedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1709500800000
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "dto.PointBrief": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INV_001_P"
                },
                "id": {
                    "type": "string",
                    "example": "point-001"
                },
                "name": {
                    "type": "string",
                    "example": "有功功率"
                },
                "type": {
                    "type": "string",
                    "example": "yaoc"
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1709500800000
                }
            }
        },
...
        "dto.UpdateDeviceRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "逆变器设备"
                },
                "ip_address": {
                    "type": "string",
                    "example": "192.168.1.101"
                },
                "manufacturer": {
                    "type": "string",
                    "example": "华为"
                },
                "model": {
                    "type": "string",
                    "example": "SUN2000-100KTL"
                },
                "name": {
                    "type": "string",
                    "example": "1号逆变器"
                },
                "port": {
                    "type": "integer",
                    "example": 502
                },
                "protocol": {
                    "type": "string",
                    "example": "modbus"
                },
                "rated_current": {
                    "type": "number",
                    "example": 150
                },
                "rated_power": {
                    "type": "number",
                    "example": 100
                },
                "rated_voltage": {
                    "type": "number",
                    "example": 380
                },
                "serial_number": {
                    "type": "string",
                    "example": "SN123456"
                },
                "slave_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT认证令牌，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}